POCs/ADK/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "AI News Podcast Generator using Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.31.1",
    "yfinance>=1.3.0",
]
